preparado tercera entrega

+++ src/controllers/product.controller.js
@@ -3,24 +3,26 @@ import { productsModel } from '../Dao/models/mongoDB.models.js';
 import { ProductsService } from '../service/index.js';
 
 class ProductController {
-    constructor(){
+    constructor() {
         this.productService = ProductsService
     }
 
     getProducts = async (req, res) => {
         try {
-            const { numPage = 1, limit = 10, sort, category} = req.query;
-    
+            console.log(req.user);
+
+            const { numPage = 1, limit = 10, sort, category } = req.query;
+
             let queryOptions = { limit, page: numPage, lean: true };
             if (sort) {
                 queryOptions.sort = { precio: parseInt(sort) };
             }
-    
+
             const { docs, page, hasNextPage, hasPrevPage, nextPage, prevPage, totalPages } = await productsModel.paginate({}, queryOptions);
-    
+
             //Productos con filtro category
             const productosCategory = docs.filter(p => p.category === category);
-    
+
             let responsePayloadCategory = {
                 productosCategory,
                 page,
@@ -28,17 +30,17 @@ class ProductController {
                 hasPrevPage,
                 totalPages
             };
-    
+
             if (hasNextPage) {
                 responsePayloadCategory.nextLink = `http://localhost:8080/api/products?numPage=${nextPage}`;
             }
-    
+
             if (hasPrevPage) {
                 responsePayloadCategory.prevLink = `http://localhost:8080/api/products?numPage=${prevPage}`;
             }
-    
-            if(category) return res.send({status: 'success', payload: responsePayloadCategory});
-    
+
+            if (category) return res.send({ status: 'success', payload: responsePayloadCategory });
+
             //Productos sin filtro category
             let responsePayload = {
                 productos: docs,
@@ -47,20 +49,20 @@ class ProductController {
                 hasPrevPage,
                 totalPages
             };
-    
+
             if (hasNextPage) {
                 responsePayload.nextLink = `http://localhost:8080/api/products?numPage=${nextPage}`;
             }
-    
+
             if (hasPrevPage) {
                 responsePayload.prevLink = `http://localhost:8080/api/products?numPage=${prevPage}`;
             }
-    
+
             return res.send({
-                status: 'success', 
+                status: 'success',
                 payload: responsePayload
             });
-    
+
         } catch (error) {
             console.log(error);
             res.status(500).send('Error interno del servidor');
@@ -69,31 +71,33 @@ class ProductController {
 
     getFilteredProducts = async (req, res) => {
         const id = req.params.pid;
-    
+
         const productosFiltradosId = await this.productService.get({ _id: id });
-    
+
         res.send(productosFiltradosId);
     }
 
     createProduct = async (req, res) => {
         try {
-            const { title, descripcion, precio, img, code, stock, category } = req.body
+            if(req.user.role === 'admin') {
+                const { title, descripcion, precio, img, code, stock, category } = req.body
+
+                const existeProducto = await this.productService.get({ code });
     
-            const existeProducto = await this.productService.get({code});
+                if (!title || !descripcion || !precio || !img || !code || !stock || !category) {
+                    console.log('Por favor, complete todos los campos para agregar producto');
+                    return res.send({ status: 'error', error: 'faltan campos' });
+                }
+                else if (existeProducto) {
+                    console.log('Los productos no pueden compartir el code');
+                    return res.send({ status: 'error', error: 'los productos no pueden compartir el code' });
+                }
     
-            if (!title || !descripcion || !precio || !img || !code || !stock || !category) {
-                console.log('Por favor, complete todos los campos para agregar producto');
-                return res.send({ status: 'error', error: 'faltan campos' });
-            }
-            else if (existeProducto) {
-                console.log('Los productos no pueden compartir el code');
-                return res.send({ status: 'error', error: 'los productos no pueden compartir el code' });
-            }
+                const productoAgregado = await this.productService.create(req.body);
     
-            const productoAgregado = await this.productService.create(req.body);
-            
-            res.status(200).send({ status: 'success', payload: productoAgregado });
-    
+                res.status(200).send({ status: 'success', payload: productoAgregado });
+            }  else return res.send('Usted debe de estar logueado y ser administrador para realizar esta tarea');
+
         } catch (error) {
             console.log(error);
             res.status(500).send({ error: 'Error interno del servidor.' });
@@ -101,32 +105,36 @@ class ProductController {
     }
 
     updateProduct = async (req, res) => {
-        try {
-            
-            const id = req.params.pid;
-            const { title, descripcion, precio, img, code, stock, category } = req.body
-    
-            if (!title || !descripcion || !precio || !img || !code || !stock || !category) {
-                console.log('Por favor, complete todos los campos para actualizar');
-                return res.send({ status: 'error', error: 'faltan campos' });
-            } else {
-                const productoActualizado = await this.productService.update({ _id: id }, req.body);
-                console.log(productoActualizado);
-                res.status(200).send({ status: 'success', payload: productoActualizado });
+        if (req.user.role === 'admin') {
+            try {
+
+                const id = req.params.pid;
+                const { title, descripcion, precio, img, code, stock, category } = req.body
+
+                if (!title || !descripcion || !precio || !img || !code || !stock || !category) {
+                    console.log('Por favor, complete todos los campos para actualizar');
+                    return res.send({ status: 'error', error: 'faltan campos' });
+                } else {
+                    const productoActualizado = await this.productService.update({ _id: id }, req.body);
+                    console.log(productoActualizado);
+                    res.status(200).send({ status: 'success', payload: productoActualizado });
+                }
+
+            } catch (error) {
+                console.log(error);
+                return res.send({ status: 'error', error: 'Not found' });
             }
-            
-        } catch (error) {
-            console.log(error);
-            return res.send({ status: 'error', error: 'Not found' });
-        }
+        } else return res.send('Usted debe de estar logueado y ser administrador para realizar esta tarea');
     }
 
     deleteProduct = async (req, res) => {
-        const id = req.params.pid;
-    
-        const productoEliminado = await this.productService.delete({ _id: id })
-    
-        res.status(200).send({ status: 'success', payload: productoEliminado });
+        if (req.user.role === 'admin') {
+            const id = req.params.pid;
+
+            const productoEliminado = await this.productService.delete({ _id: id });
+
+            res.status(200).send({ status: 'success', payload: productoEliminado });
+        } else return res.send('Usted debe de estar logueado y ser administrador para realizar esta tarea');
     }
 }
